fix(plugins/include-code): retrieve sources inside doc-api, part 2
Inside doc-api the `root` variable points not to the doc root, but to
the cache root. So to build the correct path to included source file
which can be retrieved from S3, we have to include prepend the actual
doc root to the absolute `includePath`, then remove `root` (not
`docRoot`!) from the `fullIncludePath` so that we get file name
relative to the cache root, which can be successfully retrieved from
S3.