fix: pop-up definition overlaps header (#457)
The pop-up of the terms overlapped the header. Previously, changes were made to @diplodoc/client to solve the problem, with ref: https://github.com/diplodoc-platform/client/pull/51

The changes caused new bug and it was decided to fix the initial bug in another way by changing the z-index of the terms themselves.